requirements.txt
djangorestframework
coreapi
hypothesis
hypothesis[django]